requirements.txt
numpy>=1.22.0
opencv-python>=4.5.5
torch>=1.13.0
torchvision>=0.14.0
PyYAML>=6.0
matplotlib>=3.5.0
ultralytics>=8.0.0  # 用于加载YOLOv11模型
Pillow>=9.0.0  # 用于中文文字渲染
PyQt5>=5.15.0  # GUI界面支持 